Replace test_fork_missing_predecessor with _block_
The BlockValidator, with a little help from the Completer, guarantees
that blocks are validated in increasing order of block num on a given
fork.

The fork test case is invalid because it assumes the BlockValidator is
trying to validate blocks out of order. This commit updates the test to
only test immediate predecessors that are missing.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>