Add tests for receiving blocks out of order
The intent of the completer is to guarantee blocks are received in increasing
order of block height on a given fork and ordinarily the chain controller and
block validator should not have to deal with this case. However, these tests
ensure that if the completer malfunctions, the chain controller and block
validator will not blow up.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>