Add StateReader trait
Separate out the get, contains and leaves functions (i.e. the read-only
functions) into the StateReader trait.  This creates sharp delineation
along read and write lines.

Additionally, makes MerkleDatabase derive clone, which simplifies the
use of the iterators in relation to the trait.  This requires a rather
minimal copy of values, since the LmdbDatabase is already a cloneable
construct.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>