Reinitialize block on failed block finalize
Block finalization could fail because of an invalid batch or missing
dependency. When this happened, None was returned in place of a block which
caused problems when trying to publish. This fixes the problem by
reinitializing the block when finalization fails and telling the engine the
block was not ready.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>