Implement ChainController in Rust
This is effectively a 1-to-1 rewrite of the Python ChainController
implementation.  It manages the current multi-threaded access to its
internal state via a RwLock around an internal state struct.

The ChainController is cloneable, which allows it to be passed between
the background thread and the main thread that launches the process.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>