Add StatePruneManager
Add a StatePruneManager to hold (and later prune) state root hashes that
are either from abandoned chains or old blocks in the chain.

This commit adds the queue which will be used as the source of state
root hashes to be pruned, and populates it via the update_queue
function.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>