Add decimate to StatePruneManager
The decimate function will take the first 10 percent of the state root
queue and attempt to prune those roots from the database.  If a root is
prune, it updates a metric.  If not, the root is pushed back on to the
queue, as presumably it is the root of a fork and would require its
successors to be pruned as well.

(`decimate` here is being used in the original Latin meaning of the
word)

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>