Remove the Sync trait bound from Engine
This trait bound was an artificial constraint imposed because state was being
shared to signal shutdown. By converting shutdown signaling to use channels, no
shared state is stored in the Engine or driver and the Sync bound can be
removed.

This especially makes Engine development easier for implementations based on
some library, where it may be difficult to make the library thread safe.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>