Fix Python CSDK driver bug
There was a bug in the Python CSDK driver that would cause messages to be lost
if the future timed out. A new future would be created but the message would be
resolved in the old future. Instead of creating a new future, hold onto the
old future and poll it until it completes.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>