Send back previous_block_id of initialized block
When initializing a block, the PoET oracle fetches chain head, attempts
to initialize the block, before sending the result of a successful
initialization back to the validator.  Previously, it would not send the
previous block id with the message, which would cause the validator to
query it again.

This includes the same block id used to initialize the block from the
original query.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>