Make BlockManager.persist atomic
The persist operation of the BlockManager should be an atomic one.

This requires that the Branch- and BranchDiffIterators have a "Sync"
version, which operates within the confines of a single lock.  These
Sync versions are not publicly accessible from the blockmanager, only
for internal use in these atomic situations.

External users of the iterators will continue to be provided the
thread-safe variants.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>