Update fork tip reference after validation
Only update the references to a tip of a fork in the fork cache after
the block has been successfully validated.  Tracking forks that are made
up of transactionally invalid blocks, requires extra references being
made by the block scheduler if a block is added that has a predecessor
that has not yet been validated.  This reference is unaccounted for
later.

By moving the fork reference update to after a block has been validated,
the block scheduler does not need to manage references to blocks at all.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>